Decompose deeplink integration (#757)
**Background**

Right now we can't handle deeplink in Decompose

**Changes**

Every component that can work with deeplinks has one:
- An argument in the constructor
- The `handleDeeplink` method that can handle deeplinks at runtime.

There are three ways to enter deeplinks:
- When starting a rooted Activity a deeplink is passed in, if the
`onCreate` intent already has a deeplink (this happens when we start the
Activity of a dead application via a deeplink),
[there](https://github.com/flipperdevices/Flipper-Android-App/blob/feature/decompose_v10/components/singleactivity/impl/src/main/java/com/flipperdevices/singleactivity/impl/SingleActivity.kt#L72)
- The `onNewIntent` method calls `handleDeeplink` on the root component,
[there](https://github.com/flipperdevices/Flipper-Android-App/blob/feature/decompose_v10/components/singleactivity/impl/src/main/java/com/flipperdevices/singleactivity/impl/SingleActivity.kt#L95-L106).
- And the worst thing is that to avoid throwing the root component
everywhere to handle internal deeplinks, you can call `handleDeeplink`
from
[LocalDeeplinkHandler](https://github.com/flipperdevices/Flipper-Android-App/blob/feature/decompose_v10/components/rootscreen/api/src/main/kotlin/com/flipperdevices/rootscreen/api/RootDeeplinkHandler.kt).
It's a very dirty way, but I thought I was willing to do it.

In the case of the argument in the constructor everything is simple and
easy - in the `initialStack` argument of the decompose we pass the stack
we need, which should be opened by the deeplink.
But in the case of opening `handleDeeplink` in rantime everything is
more complicated - first I check if the child component I want to access
is alive. If it is alive, I call `handleDeeplink` already from it and
further down the hierarchy + plus I bring this component to the top with
the same config stored in the stack to avoid recreating the component.
If it is not alive, then I just put up a new config with the deeplink
tucked in. It looks like this [like
this](https://github.com/flipperdevices/Flipper-Android-App/blob/feature/decompose_v10/components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/api/BottomBarDecomposeComponentImpl.kt#L171-L205)

**Test plan**

Try open all deeplinks